LayoutTests/fast/dom/Window/window-appendages-cleared-expected.txt

PASS history.length == "LEFTOVER" is false
PASS history.back == "LEFTOVER" is false
PASS history.go == "LEFTOVER" is false
PASS history.forward == "LEFTOVER" is false
PASS frames.location == "LEFTOVER" is false
PASS frames.length == "LEFTOVER" is false
PASS location.host == "LEFTOVER" is false
PASS location.pathname == "LEFTOVER" is false
PASS location.port == "LEFTOVER" is false
PASS location.href == "LEFTOVER" is false
PASS location.search == "LEFTOVER" is false
PASS location.assign == "LEFTOVER" is false
PASS location.reload == "LEFTOVER" is false
PASS location.hash == "LEFTOVER" is false
PASS location.hostname == "LEFTOVER" is false
PASS location.protocol == "LEFTOVER" is false
PASS location.toString == "LEFTOVER" is false
PASS location.replace == "LEFTOVER" is false
PASS locationbar.visible == "LEFTOVER" is false
PASS menubar.visible == "LEFTOVER" is false
PASS personalbar.visible == "LEFTOVER" is false
PASS scrollbars.visible == "LEFTOVER" is false
PASS statusbar.visible == "LEFTOVER" is false
PASS toolbar.visible == "LEFTOVER" is false
PASS window.event is undefined.
PASS successfullyParsed is true

TEST COMPLETE
